django/QuizApp/requirements.txt
asgiref==3.3.1
Django==3.1.6
django-adminactions==1.9
pytz==2021.1
sqlparse==0.4.1
xlrd==2.0.1
xlwt==1.3.0
